Role based Authentication
added different with admin controller

+++ src/main/java/com/ferdeen/Journal/Application/Service/JournalEntryService.java
@@ -51,11 +51,23 @@ public class JournalEntryService {
         return journalEntryRepo.findById(myid);
     }
 
-    public void deletebyId(ObjectId myid, String userName) {
-     User user =   userService.findbyUserName(userName);
-     user.getJournalEntries().removeIf(x -> x.getId().equals(myid));
-     userService.saveUser(user);
-        journalEntryRepo.deleteById(myid);
+    @Transactional
+    public boolean deletebyId(ObjectId myid, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findbyUserName(userName);
+             removed = user.getJournalEntries().removeIf(x -> x.getId().equals(myid));
+            if (removed) {
+                userService.saveUser(user);
+                journalEntryRepo.deleteById(myid);
+            }
+        } catch (Exception e){
+            System.out.println(e);
+            throw  new RuntimeException("An error occured while saving the entry" , e);
+        }
+
+        return removed;
+
     }
 
 }
